Updated and expanded cloud mask functions and added basic tests
Added separate cloud mask functions for Landsat SR and Sentinel 2.  This will need additional work to identify which bits to use and whether to apply additional buffering (and/or shrinking) to the masks.  Currently snow pixels are being masked although the EEFlux team mentioned this may not be ideal.

One thing to note is that the masks are currently being built so that clouds are 0 and clear is 1 so that they can be used directly by an updateMask call, but it might make more sense to flip this back.  This would be easy to do and would only requre removin the final .Not() call.