Moved all code into a "core" folder/module
I was having problems getting travis to run on the models with everything in the root of the openet folder.  I'm not totally sure why but it seems like having it all in a separate folder seems to work.  For now I'm not changing the PyPi name but this might be needed also.