Switched interpolate daily to not multiple target and source images
This will probably get switched back at some point, but I think it is simpler for now to simply return both the target and source images instead of trying to scale or multiply them together.  Once we have a better sense of how the different models might use the interpolation functions we can change this.

Also, changed the utils image value extract functions to return dictionaries intead of values or list of values.