Updated daily interpolator code
Removed the saveAll joins and instead am building the prev/next collections inside the loop.  The drawback to this approach is that the "linear" function is now buried inside the daily function and is not being tested separately.  The benefit of this approach is that I have much better control over the interpolation for short interpolation windows and interpolations past the last known value now hold a constant value.

In order to minimize the number of map calls, I am not adding the time band inside the interpolation and am instead requiring that it be included as one of the target variables.  This means the models must support writing a time band.  This may get changed at some point if it doesn't seem like this is saving any time.

All the time operations are currently being done for 0 UTC dates.