Added a commented out test for checking the use_joins approach for date ranges with missing data
The use_joins approach doesn't work well with small interpolation ranges when there are not images available because the saveAll join drops images that don't have any matching iamges.

Also, I started changing the test images a little to make it easier to switch to using floats for the interpolation.