Updating interpolation tests
Checking eto and etr values for many of the tests.  This isn't really needed but it is good to check both in case the GRIDMET data changed in one band.

Changed the scale parameter to 30.  This does change some of the test values but it seems better to be consistent with the Landsat scale.