Updated point image and point collection value tests
Calls to the USGS/NED asset were returning None for some versions of the EE api.  The asset may have been changed or is being deprecated.  Switched the tests to check the NASA DEM and USGS 3DEP assets instead.

Also, renamed tests so that utils and landsat tests run first.  There is probably a correct way of doing this in pytest.