Adding a generic normalized difference index function
This function can be used to computed the normalized difference index for any combination of bands with support for masking saturated and negative reflectance pixels, and allow the user to set a water mask so that negative or low reflectance pixels can be set to a different output value (e.g. -1 instead of 0).  The original NDVI function now calls this function but with the correct band names.